Add line-of-sight functionality to Enemy vision
Updated EnemyAI script
Added CheckLineOfSight(), which determines whether the player is seen or not. The player first gets noticed when collides with enemy's VisionCone - we get player's object in OnTriggerEnter(). Then enemy follows  player based on CheckLineOfSight(). The player escapes if gets out of the line of sight (there is an obstacle between player and enemy) or line of sight's range (player goes too far from the enemy). If player gets out, enemy forgets about him and returns back to his patrol route.